Add Java-style "for" loop:
for (INIT [, INIT]* ; CONDITION ; ITERATE [, ITERATE]* ) BODY

You can have 0 or more INIT statements, which can be of the form
X = EXP (where X is declared already) or
TYPE X = EXP (whre X of type TYPE will be declared for use in the BODY scope).

You can have 0 or more ITERATE statements, which can be of the form
X++ or X--
++X or -- X
X OPER EXP (where OPER can be =, +=, -=, etc.

BODY can be a single statement ending with a ; or a block enclosed in {}, with
no ; at the end, just like and other loop.

break, continue, return, or exit are allowed within BODY, just like any loop.


git-svn-id: https://svn.code.sf.net/p/kolmafia/code@17460 b29ace70-8910-0410-8dcc-aa2fc6433167